+++ services/services/player/ipc/player_service_proxy.h
@@ -45,7 +45,7 @@ public:
     int32_t SetSourceLoader(const sptr<IRemoteObject> &object) override;
     int32_t SetMediaSource(const std::shared_ptr<AVMediaSource> &mediaSource, AVPlayStrategy strategy) override;
     int32_t AddSubSource(int32_t fd, int64_t offset, int64_t size) override;
-    int32_t Seek(int32_t mSeconds, PlayerSeekMode mode) override;
+    int32_t Seek(int64_t mSeconds, PlayerSeekMode mode) override;
     int32_t SeekToDefaultPosition() override;
     int32_t Stop() override;
     int32_t GetPlaybackSpeed(PlaybackRateMode &mode) override;
